+++ core/src/consensus/fork_choice.rs
@@ -1,17 +1,35 @@
 use {
-    crate::{
-        consensus::{
-            latest_validator_votes_for_frozen_banks::LatestValidatorVotesForFrozenBanks,
-            progress_map::ProgressMap, SwitchForkDecision, Tower,
-        },
-        replay_stage::HeaviestForkFailures,
+    crate::consensus::{
+        latest_validator_votes_for_frozen_banks::LatestValidatorVotesForFrozenBanks,
+        progress_map::ProgressMap, SwitchForkDecision, Tower,
     },
     solana_runtime::{bank::Bank, bank_forks::BankForks},
+    solana_sdk::slot_history::Slot,
     std::{
         collections::{HashMap, HashSet},
         sync::{Arc, RwLock},
     },
 };
+
+#[derive(PartialEq, Eq, Debug)]
+pub enum HeaviestForkFailures {
+    LockedOut(u64),
+    FailedThreshold(
+        Slot,
+        /* Observed stake */ u64,
+        /* Total stake */ u64,
+    ),
+    FailedSwitchThreshold(
+        Slot,
+        /* Observed stake */ u64,
+        /* Total stake */ u64,
+    ),
+    NoPropagatedConfirmation(
+        Slot,
+        /* Observed stake */ u64,
+        /* Total stake */ u64,
+    ),
+}
 
 pub struct SelectVoteAndResetForkResult {
     pub vote_bank: Option<(Arc<Bank>, SwitchForkDecision)>,
