+++ core/src/repair/repair_service.rs
@@ -8,7 +8,7 @@ use {
 use {
     crate::{
         cluster_info_vote_listener::VerifiedVoteReceiver,
-        cluster_slots_service::cluster_slots::ClusterSlots,
+        consensus::cluster_slots::ClusterSlots,
         repair::{
             ancestor_hashes_service::{AncestorHashesReplayUpdateReceiver, AncestorHashesService},
             duplicate_repair_status::AncestorDuplicateSlotToRepair,
